% Auto-generated: do not edit by hand
\name{sTOMPWebsocket}

\alias{sTOMPWebsocket}

\title{STOMPWebsocket component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
sTOMPWebsocket(id=NULL, message=NULL, send=NULL, subscribe=NULL,
unsubscribe=NULL, url=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{message}{Named list. The message from subscription.}

\item{send}{Named list. The message to send}

\item{subscribe}{Character. The topic to subscribe to.}

\item{unsubscribe}{Character. The topic to unsubscribe from.}

\item{url}{Character. The url to connect to}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

